feat: Uniform Random Rotation + Tests
TODO: implement Uniform Random Rotation in theano.tensor + test that it produce an uniform distribution in coord ussing Poisson.

Uniform Random Rotation work good with numpy. The show_dist_quaternions functions show the distribution in coords of the result of rotate a unitary vector 8*600 times

Unitary Test added to check if rotation still produce unitary vector, and an intuitive way of test that the distribution is uniform.